Rename `MetronomeContext` to `Metronome`
