+++ src/index.tsx
@@ -3,7 +3,7 @@ import ReactDOM from 'react-dom/client'
 import App from './App'
 import reportWebVitals from './reportWebVitals'
 import './index.css'
-import { MetronomeProvider } from './MetronomeContext'
+import { MetronomeProvider } from './Metronome'
 import { AudioProvider } from './AudioRouter'
 
 const root = ReactDOM.createRoot(document.getElementById('root') as HTMLElement)
